Rename: 엔티티 이름 변경(EmailNotification, KakaoTalkNotification)
